cleaning dev a bit to align with IoW needs and general use
